graphql_api.py: Raise an exception for unexpected HTTP error codes
Occasional HTTP server error codes can be returned from the GraphQL
endpoint, in which case the workaround is to retry the same query again.
The HTTP status code is placed in the exception object along with an
error message in this case, so that callers can get more information
other than the KeyError exception that is likely to happen when assuming
there is data returned from r.json(), and enables retries for HTTP 502
and 503 error codes.