Collection: add option to include media files when downloading
New 'include_media' option is off by default, for consistency with
corresponding options for metadata and associated files. When enabled,
it will download any images and/or videos in the collection.

Note that the 'delete' flag is currently ignored by the helper method
for downloading media, except for printing out a "NYI" warning if
'delete' is enabled. Deleting local media files is not as convenient
as deleting metadata files, since they are not segregated out into their
own little subfolder.