fix naming of metadata files
Videos have both "name" and "filename" attributes in Conservator.
It is the "filename" attribute that can be expected to have a file suffix
according to filetype (for instance ".mp4" or ".avi"), so that should
work with the current scheme for naming downloaded metadata after the
corresponding video:

  some-video.mp4 -> video_metadata/some-video.json
  other-video.mp4 -> video_metadata/other-video.json

The video "name" defaults to same as "filename" so using "name" worked
for projects that left the default names alone, but the result was not so
useful for projects that gave videos non-default names that no longer
included a suffix:

  some-video -> video_metadata/json
  other-video -> video_metadata/json

meaning, the same "video_metadata/json" file kept getting overwritten,
ending up with just one set of metadata actually downloaded (the last
one processed).

Thus, metadata filename will now be generated based on the "filename"
attribute of the corresponding video, not on its "name" attribute.