graphql_api.py: implement retries for transient errors
There are a number of transient errors that can occur in graphql connections,
both due to server intentionally throttling response rate and due to network
glitches.

Add a configurable number of retries (max_retries parameter, set to
10 by default) with a configurable delay between them (retry_delay
parameter, set to 1.0 sec by default).

An exception will still be thrown if all the retries fail, and in that case the
exception message will be set according to how the last attempt went.
Errors returned by a well-formed graphql response message are
NOT considered transient errors; they are assumed to be a problem in the
construction of the query, and cause an exception without bothering to
attempt retries.