graphql_api.py: fix get_*_filelist() functions
get_{video,image,media}_filelist() functions were depending on getting
correct media counts from the collection to calculate the number of
queries to make (they are based on paged queries instead of one big one),
but there is currently a conservator bug in non-recursive media counts.

For example, current results for YYJ_Video_Analytics folder show:
{'recursiveVideoCount': 1532, 'videoCount': 1000, 'recursiveImageCount': 0, 'imageCount': 0}
The current count in this folder is actually 1460, not 1000.

ConservatorTools repo has a workaround for the bug, but it is simpler
to just keep making queries until hitting a page with zero results
(or an error).