Consolidate media creation in MediaType
These mutations may be named after "Video", but they are actually used
for both videos and images alike. They don't belong in a video-specific
class, but rather in the class common to both, MediaType.

Since upload() is the way to create a new Video or Image object in
conservator, it also makes sense for it to be implemented in the
MediaType manager rather than in Video. Also like the symmetry of upload()
and download() being in the same class.

Also renamed the intermediate steps of uploading a media file to be
private methods, since it seems very rare that outside code would
actually need to call any of those individual steps from any other
context. For example, it didn't really make sense for a collection
object to be able to call "create_video", which just allocates a
placeholder id without being an actual useful video until the whole
upload process has been completed.