Add verbose options for methods that run git commands
For task_runner, it is cleaner to keep stdout and stderr output to a
minimum.  Some git commands (e.g. clone and pull) display output that
can be disabled using the -q option.  A new `verbose` parameter in
the method calls controls whether -q option is used when calling git
commands that output to the terminal:
* verbose=True skips -q (current behavior)
* verbose=False passes in -q

The push_commits() method has been enhanced with a check of the output
to verify that the index.json changes were accepted.  If not, a warning
message is generated if git responds with "Everything up-to-date",
otherwise an exception will be generated.