Add retries to clone of datasets
Account for the race condition when a dataset has just been created using an
API call and its repository is not immediately available.

When working with a dataset that was already fully initialized the first
attempt to clone it will succeed, so there will be no sleep, but
if the dataset repo is still being initialized then there will be
N retries spaced apart M seconds. Default is 2 retries with 5 sec wait
between, giving the repo initializatioon up to 10 sec to complete.