optimize downloads by skipping already-downloaded files
This is analogous to the upload --resume_media option, but is
enabled all the time instead of being optional since the overhead
is so low. We're already querying every media, so just including the
md5 (or image_md5) field is almost for free. Then we have the md5sum
handy in case there is an existing file at the destination local path,
and decide whether that local file is valid or not.

download_file() was changed to take md5 argument for when it is available;
when it is not available, just put in an empty string. Don't want it
to be an arg with default, because sometimes passing it and sometimes
not could mess with the way download_files() is passing it arguments.