Fix project name error when creating a new remote collection path
If a project doesn't exist, create_from_remote_path() would produce a bad
GraphQL query with a '/' in front of the project name.  Generally, new
projects won't need to be created, but test code will be likely to be
creating new projects on a local development Conservator server, so it
helps if this works correctly.