CLI collections create-path command: Rename -r to -p
As this command most resembles 'mkdir' for creating new paths, rename the
'-r' (recursive) option to '-p' (parents) for consistency.  Keeping this
option is helpful to avoid creating new paths by mistake in case of typos,
when the base path is expected to be present already.

Clean up conditionals as recommended in PR review.