fix kubernetes ingress failure in jenkins pipeline
jenkins test started to fail with this complaint:

  Error from server (InternalError): error when creating "STDIN": Internal
    error occurred: failed calling webhook "validate.nginx.ingress.kubernetes.io":
    an error on the server ("") has prevented the request from succeeding

Turns out the jenkins test was using an ingress config file from master
branch of an upstream repo, *not* locked to a specific version.

The copy that works with our current 0.10 version of KinD definitly
has some differences from the copy in the master branch now,
for instance the working one said
    app.kubernetes.io/version: 0.44.0
where master now says
    app.kubernetes.io/version: 0.47.0

Switch to using the older version of this config.