integration test test_collections.py expected failure work-around
A recent bug in flirconservator returns an error when removing a media
file from a collection if at creation time the media file specified the
collection's ID.  To work around this error, test_remove_media() was
marked as an expected failure, and a replacement method
test_remove_media_workaround() was added that does not specify the
collection ID when uploading the media file, but instead adds the media
file to the collection before removal.

Assuming the bug is fixed, the original test method can have the
@pytest.mark.xfail() decorator removed, and the workaround method can
be deleted.