Update package version handling to unbreak python packaging
The pypi release process has started using a newer python 'build'
module which now tries to build the python package out-of-tree,
so the git version is no longer available to look up.

Fix this by making sure the git version is cached into a file
that can be packed up with the other source files. Reading the
git version is preferred when building from the git repo,
but in other cases (like packaging for pypi) the cached version
will be used.