integration tests: update kubernetes port-forward for mongodb
Previously, mongodb running inside kubernetes was mapped to a different port on
localhost than its standard port, presumably to allow a mongodb server to
run in the host at the same time without conflict.

That strategy of mapping mongodb to different port on localhost does not work
if the mongodb instance used for testing has replica set enabled. Then there
would be a mismatch between the port in replica set config and the port where
mongodb is actually available.

Note that there is one other caveat for using replic sets with mongodb in a
kubernets cluster: you also need to be able to resolve whatever hostname
was used inside the replic set config (see 'rs.config()' in mongo shell)
as 'localhost' so that it will end up at that forwarded port. That can
be done with a suitable /etc/hosts entry, for instance

  $ grep mongo /etc/hosts
  127.0.0.1       conservator-mongo