start-cluster.sh: allow overriding kubernetes dir instead of run.sh
Now start-cluster.sh takes an optional argument, which is a path
to a directory containing kubernetes config directory for bringing
up a conservator instance. If the path is not supplied, then the
kubernetes config directory will be pulled from a docker image
(like it was before).

The mechanism to override just kubernetes/kind/run.sh is being
removed, since now you can pass in a modified kubernetes/ dir instead.