address race condition in cvc publish
'cvc publish' was sometimes hitting an error state when trying
to sync the local directory with the server after pushing changes.
It would print error messages and the local directory would be stale,
rather than in a position to do further 'publish' commands:

  remote: error: Could not read 3e129193b6dfc84392152fad0fa4714f249eeb99
  remote: fatal: bad tree object 3e129193b6dfc84392152fad0fa4714f249eeb99
  remote: aborting due to possible repository corruption on the remote side.
  fatal: protocol error: bad pack header
  branch 'master' set up to track 'origin/master'.
  Reset branch 'master'
  Your branch is up to date with 'origin/master'.

It seems that the 'pull' after 'push' done Dataset.commit() was
sometimes coming while the new commit was still being processed on
the server.

Avoid this problem via a new Dataset method wait_for_history_len()
that waits for:
* commit log to contain expected number of commits (at least one
  more than before the 'push')
* dataset's git_commit_state field to be "completed" rather than
  "processing"

There is a max time to wait in case git processing gets wedged
on the server side; if the timeout is reashed, Dataset.commit()
will issue a warning message instead of trying to do the 'pull'.