validate each field when creating new config
Instead of waiting for the full config to be entered and making the
user start from beginning if one thing is wrong, validate the
config after each field is input.

The API key needs to be after url in the list of fields,
because it can only be checked by trying to do a query against
the url. Made a note that API key should stay after url in the list.