Implement methods to add/update dataset annotations
Also totally rewrite the example script for updating dataset metadata,
because the single-file-upload method of changing dataset metadata is quite
deprecated. Now it will do a separate API call to update every annotation that
needs changed. The update includes QA fields, so it will be just one call for
any changes to an annotation (aside from attributes, which are not supported
in the example script).