Add delays when creating or deleting datasets in integration tests
The server needs time to commit a new dataset (or to commit a dataset
before deleting it), so give extra time before failing integration tests
that involve dataset creation or deletion.

Also, when checking for dataset status, look for its ID in the list of
all search results instead of assuming that the count of search results
provides enough information.